ml_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy>=1.26.2
scikit-learn==1.4.0
scikit-surprise==1.1.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
joblib==1.3.2
